remove outdated files as well as outdated trend calcs
