+++ requirements.txt
@@ -7,3 +7,4 @@ scipy
 pytest
 pytest-profiling
 graphviz
+statsmodels.api
