--- top_n_bottom.json
@@ -1,227 +1,0 @@
-[
-  {
-    "name": "Bjuv",
-    "region": "Skåne län",
-    "emissions": {
-      "2015": 70032.85,
-      "2016": 68342.36,
-      "2017": 55939.57,
-      "2018": 50425.9,
-      "2019": 50027.65,
-      "2020": 47925.45,
-      "2021": 48362.26,
-      "2022": 40743.65,
-      "2023": 27539.54
-    }
-  },
-  {
-    "name": "Härjedalen",
-    "region": "Jämtlands län",
-    "emissions": {
-      "2015": 84787.05,
-      "2016": 80862.78,
-      "2017": 72787.01,
-      "2018": 40243.31,
-      "2019": 42647.62,
-      "2020": 40412.1,
-      "2021": 38004.94,
-      "2022": 33556.55,
-      "2023": 31934.83
-    }
-  },
-  {
-    "name": "Stockholm",
-    "region": "Stockholms län",
-    "emissions": {
-      "2015": 1796962.18,
-      "2016": 1659530.13,
-      "2017": 1531684.14,
-      "2018": 1520265.14,
-      "2019": 1270604.57,
-      "2020": 920146.4,
-      "2021": 999825.3,
-      "2022": 933443.26,
-      "2023": 874782.94
-    }
-  },
-  {
-    "name": "Åre",
-    "region": "Jämtlands län",
-    "emissions": {
-      "2015": 48863.63,
-      "2016": 47245.44,
-      "2017": 45725.31,
-      "2018": 27447.13,
-      "2019": 27999.17,
-      "2020": 25868.89,
-      "2021": 25443.39,
-      "2022": 23015.43,
-      "2023": 22915.89
-    }
-  },
-  {
-    "name": "Berg",
-    "region": "Jämtlands län",
-    "emissions": {
-      "2015": 49862.87,
-      "2016": 48773.95,
-      "2017": 42986.9,
-      "2018": 28990.67,
-      "2019": 28607.0,
-      "2020": 27202.85,
-      "2021": 26887.98,
-      "2022": 23940.41,
-      "2023": 24442.24
-    }
-  },
-  {
-    "name": "Upplands-Bro",
-    "region": "Stockholms län",
-    "emissions": {
-      "2015": 77526.21,
-      "2016": 75219.07,
-      "2017": 90626.01,
-      "2018": 82262.4,
-      "2019": 81330.72,
-      "2020": 172343.33,
-      "2021": 182432.55,
-      "2022": 197017.58,
-      "2023": 190911.77
-    }
-  },
-  {
-    "name": "Pajala",
-    "region": "Norrbottens län",
-    "emissions": {
-      "2015": 33826.74,
-      "2016": 30985.16,
-      "2017": 32854.37,
-      "2018": 32932.11,
-      "2019": 44794.56,
-      "2020": 56531.01,
-      "2021": 55534.91,
-      "2022": 69278.67,
-      "2023": 72856.8
-    }
-  },
-  {
-    "name": "Timrå",
-    "region": "Västernorrlands län",
-    "emissions": {
-      "2015": 73486.52,
-      "2016": 87326.95,
-      "2017": 92346.95,
-      "2018": 125641.95,
-      "2019": 104718.47,
-      "2020": 97789.81,
-      "2021": 80675.8,
-      "2022": 112864.66,
-      "2023": 104108.66
-    }
-  },
-  {
-    "name": "Karlshamn",
-    "region": "Blekinge län",
-    "emissions": {
-      "2015": 110149.4,
-      "2016": 101749.59,
-      "2017": 87198.14,
-      "2018": 113339.51,
-      "2019": 84923.62,
-      "2020": 87152.14,
-      "2021": 160465.31,
-      "2022": 237732.04,
-      "2023": 91582.04
-    }
-  },
-  {
-    "name": "Munkfors",
-    "region": "Värmlands län",
-    "emissions": {
-      "2015": 8178.59,
-      "2016": 7800.53,
-      "2017": 7060.32,
-      "2018": 6480.22,
-      "2019": 6138.84,
-      "2020": 6217.41,
-      "2021": 6121.41,
-      "2022": 5479.38,
-      "2023": 9649.51
-    }
-  },
-  {
-    "name": "Dals-Ed",
-    "region": "Västra Götalands län",
-    "emissions": {
-      "2015": 14351.96,
-      "2016": 13700.43,
-      "2017": 14423.81,
-      "2018": 13018.74,
-      "2019": 13101.17,
-      "2020": 12320.19,
-      "2021": 12443.28,
-      "2022": 10912.24,
-      "2023": 11148.3
-    }
-  },
-  {
-    "name": "Robertsfors",
-    "region": "Västerbottens län",
-    "emissions": {
-      "2015": 43694.21,
-      "2016": 42554.43,
-      "2017": 45901.05,
-      "2018": 42875.94,
-      "2019": 42074.45,
-      "2020": 36930.31,
-      "2021": 36419.25,
-      "2022": 34133.31,
-      "2023": 34019.5
-    }
-  },
-  {
-    "name": "Östhammar",
-    "region": "Uppsala län",
-    "emissions": {
-      "2015": 86130.56,
-      "2016": 85334.99,
-      "2017": 86914.02,
-      "2018": 84608.12,
-      "2019": 80773.02,
-      "2020": 67266.19,
-      "2021": 67671.74,
-      "2022": 67438.33,
-      "2023": 67007.81
-    }
-  },
-  {
-    "name": "Knivsta",
-    "region": "Uppsala län",
-    "emissions": {
-      "2015": 71910.34,
-      "2016": 69523.49,
-      "2017": 69364.81,
-      "2018": 69476.66,
-      "2019": 68090.39,
-      "2020": 60443.2,
-      "2021": 60930.3,
-      "2022": 56327.21,
-      "2023": 56379.74
-    }
-  },
-  {
-    "name": "Norrtälje",
-    "region": "Stockholms län",
-    "emissions": {
-      "2015": 178652.4,
-      "2016": 172571.87,
-      "2017": 169858.96,
-      "2018": 169809.96,
-      "2019": 164579.74,
-      "2020": 152063.38,
-      "2021": 152453.11,
-      "2022": 141520.94,
-      "2023": 140632.44
-    }
-  }
-]
